Rename module sample to mapzen-android-sdk-sample
