+++ mapzen-android-sdk-sample/.gitignore
