+++ mapzen-android-sdk-sample/build.gradle
@@ -8,7 +8,7 @@ android {
   buildToolsVersion "24.0.3"
 
   defaultConfig {
-    applicationId "com.mapzen.android.sample"
+    applicationId "com.mapzen.android.sdk.sample"
     minSdkVersion 15
     targetSdkVersion 24
     versionCode 1
