+++ mapzen-android-sdk-sample/proguard-rules.pro
