+++ mapzen-android-sdk-sample/src/main/AndroidManifest.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
-<manifest package="com.mapzen.android.sample"
+<manifest package="com.mapzen.android.sdk.sample"
     xmlns:android="http://schemas.android.com/apk/res/android">
 
   <application
