+++ mapzen-android-sdk-sample/src/main/assets/styles/tron/tron.yaml
