+++ mapzen-android-sdk-sample/src/main/java/com/mapzen/android/sdk/sample/MarkerMapzenActivity.java
@@ -1,18 +1,18 @@
-package com.mapzen.android.sample;
+package com.mapzen.android.sdk.sample;
 
 import com.mapzen.android.graphics.MapFragment;
 import com.mapzen.android.graphics.MapzenMap;
 import com.mapzen.android.graphics.OnMapReadyCallback;
-import com.mapzen.android.graphics.model.Polygon;
+import com.mapzen.android.graphics.model.Marker;
 import com.mapzen.tangram.LngLat;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 
 /**
- * Polygon SDK demo.
+ * Show map markers.
  */
-public class PolygonMapzenActivity extends AppCompatActivity {
+public class MarkerMapzenActivity extends AppCompatActivity {
 
   private MapzenMap map;
 
@@ -24,26 +24,24 @@ public class PolygonMapzenActivity extends AppCompatActivity {
         (MapFragment) getSupportFragmentManager().findFragmentById(R.id.fragment);
     mapFragment.getMapAsync(new OnMapReadyCallback() {
       @Override public void onMapReady(MapzenMap map) {
-        PolygonMapzenActivity.this.map = map;
+        MarkerMapzenActivity.this.map = map;
         configureMap();
       }
     });
   }
 
   private void configureMap() {
-    Polygon polygon = new Polygon.Builder().add(new LngLat(-73.9903, 40.74433))
-        .add(new LngLat(-73.984770, 40.734807))
-        .add(new LngLat(-73.998674, 40.732172))
-        .add(new LngLat(-73.996142, 40.741050))
-        .build();
-    map.addPolygon(polygon);
+    map.addMarker(new Marker(-73.9903, 40.74433));
+    map.addMarker(new Marker(-73.984770, 40.734807));
+    map.addMarker(new Marker(-73.998674, 40.732172));
+    map.addMarker(new Marker(-73.996142, 40.741050));
 
     map.setZoom(15f);
     map.setPosition(new LngLat(-73.9918, 40.73633));
   }
 
   @Override protected void onDestroy() {
-    map.removePolygon();
+    map.removeMarker();
     super.onDestroy();
   }
 }
