+++ mapzen-android-sdk-sample/src/main/java/com/mapzen/android/sdk/sample/PolygonMapzenActivity.java
@@ -1,18 +1,18 @@
-package com.mapzen.android.sample;
+package com.mapzen.android.sdk.sample;
 
 import com.mapzen.android.graphics.MapFragment;
 import com.mapzen.android.graphics.MapzenMap;
 import com.mapzen.android.graphics.OnMapReadyCallback;
-import com.mapzen.android.graphics.model.Polyline;
+import com.mapzen.android.graphics.model.Polygon;
 import com.mapzen.tangram.LngLat;
 
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 
 /**
- * Polyline SDK demo.
+ * Polygon SDK demo.
  */
-public class PolylineMapzenActivity extends AppCompatActivity {
+public class PolygonMapzenActivity extends AppCompatActivity {
 
   private MapzenMap map;
 
@@ -24,26 +24,26 @@ public class PolylineMapzenActivity extends AppCompatActivity {
         (MapFragment) getSupportFragmentManager().findFragmentById(R.id.fragment);
     mapFragment.getMapAsync(new OnMapReadyCallback() {
       @Override public void onMapReady(MapzenMap map) {
-        PolylineMapzenActivity.this.map = map;
+        PolygonMapzenActivity.this.map = map;
         configureMap();
       }
     });
   }
 
   private void configureMap() {
-    Polyline polyline = new Polyline.Builder().add(new LngLat(-73.9903, 40.74433))
+    Polygon polygon = new Polygon.Builder().add(new LngLat(-73.9903, 40.74433))
         .add(new LngLat(-73.984770, 40.734807))
         .add(new LngLat(-73.998674, 40.732172))
         .add(new LngLat(-73.996142, 40.741050))
         .build();
-    map.addPolyline(polyline);
+    map.addPolygon(polygon);
 
     map.setZoom(15f);
     map.setPosition(new LngLat(-73.9918, 40.73633));
   }
 
   @Override protected void onDestroy() {
-    map.removePolyline();
+    map.removePolygon();
     super.onDestroy();
   }
 }
