+++ mapzen-android-sdk-sample/src/main/java/com/mapzen/android/sdk/sample/SearchResultsActivity.java
@@ -1,4 +1,4 @@
-package com.mapzen.android.sample;
+package com.mapzen.android.sdk.sample;
 
 import com.mapzen.android.graphics.MapFragment;
 import com.mapzen.android.graphics.MapzenMap;
