+++ mapzen-android-sdk-sample/src/main/res/layout/activity_clear_btn.xml
