+++ mapzen-android-sdk-sample/src/main/res/layout/activity_main.xml
