+++ mapzen-android-sdk-sample/src/main/res/layout/activity_route.xml
