+++ mapzen-android-sdk-sample/src/main/res/layout/activity_sample_mapzen.xml
