+++ mapzen-android-sdk-sample/src/main/res/layout/activity_touch_responder.xml
