+++ mapzen-android-sdk-sample/src/main/res/layout/list_item.xml
