+++ mapzen-android-sdk-sample/src/main/res/layout/simple_spinner_dropdown_item.xml
