+++ mapzen-android-sdk-sample/src/main/res/menu/menu_sample_mapzen.xml
