+++ mapzen-android-sdk-sample/src/main/res/mipmap-hdpi/ic_launcher.png
