+++ mapzen-android-sdk-sample/src/main/res/mipmap-xxxhdpi/ic_launcher.png
