+++ mapzen-android-sdk-sample/src/main/res/values-v21/styles.xml
