+++ mapzen-android-sdk-sample/src/main/res/values/colors.xml
