+++ mapzen-android-sdk-sample/src/main/res/values/dimens.xml
