+++ mapzen-android-sdk-sample/src/main/res/values/mapzen.xml
