+++ mapzen-android-sdk-sample/src/main/res/values/strings.xml
