+++ mapzen-android-sdk-sample/src/main/res/values/styles.xml
